implemetacion de colisiones
