+++ src/GameRoutes.jsx
@@ -2,7 +2,7 @@ import { BrowserRouter, Route, Routes } from "react-router-dom";
 import Level2 from "./levels/Level2/Level2";
 import Experience from './Experience';
 import Level1 from "./levels/Level1";
-import Level6 from "./levels/Level6/Level6";
+import Level3 from "./levels/Level3/Level3";
 
 export default function GameRoutes() {
     return (
@@ -11,7 +11,7 @@ export default function GameRoutes() {
                 <Route path="/" element={<Experience />} />
                 <Route path="/level1" element={<Level1 />} />
                 <Route path="/level2" element={<Level2 />} />
-                <Route path="/level6" element={<Level6 />} />
+                <Route path="/level3" element={<Level3 />} />
                 
             </Routes>
         </BrowserRouter>
